+++ dmsr/dmsr_gan/__init__.py
@@ -3,4 +3,3 @@ from .generator import *
 from .critic import *
 from .density_critic import *
 from .monitor import *
-from .dataset import *
