+++ scripts/tests/dataset_test.py
@@ -16,14 +16,12 @@ import matplotlib.pyplot as plt
 
 from torch.utils.data import DataLoader
 
-from swift_tools.data import load_numpy_dataset
-from dmsr.dmsr_gan import DMSRDataset
+from dmsr.data_tools import load_numpy_dataset
+from dmsr.data_tools import DMSRDataset
 from dmsr.field_operations.conversion import displacements_to_positions
 
 
 #%%
-
-
 def plot_positions(positions):
     positions = torch.transpose(positions, 1, -1)
     positions = positions.reshape((-1, 3))
@@ -35,7 +33,6 @@ def plot_positions(positions):
     plt.close()
 
 
-
 data_directory = '../../data/dmsr_training_velocity_x64/'
 data = load_numpy_dataset(data_directory)
 LR_data, HR_data, box_size, LR_grid_size, HR_grid_size = data
@@ -45,7 +42,6 @@ dataset = DMSRDataset(LR_data, HR_data, augment=True)
 
 
 #%%
-
 # DataLoader to iterate over the dataset
 dataloader = DataLoader(dataset, batch_size=1, shuffle=False)
